backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
gradio==4.7.1
torch==2.2.1
torchvision==0.17.1
diffusers==0.24.0
transformers==4.35.2
accelerate==0.24.1
pillow==10.1.0
numpy==1.26.2
python-dotenv==1.0.0 